pyproject.toml
[project]
name = "test-skill"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agent-sandbox>=0.0.30",
    "deepagents>=0.5.3",
    "dotenv>=0.9.9",
    "langchain>=1.2.15",
    "langchain-litellm>=0.6.4",
    "langchain-openai>=1.1.13",
    "langgraph-cli[inmem]>=0.4.21",
]
